requirements.txt
fastprogress
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
torch==1.0.0
torchvision==0.2.1

